+++ src/main/java/com/sparta/quokkatravel/domain/admin/coupon/service/CouponProducerService.java
@@ -1,13 +1,13 @@
 package com.sparta.quokkatravel.domain.admin.coupon.service;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sparta.quokkatravel.domain.accommodation.entity.Accommodation;
 import com.sparta.quokkatravel.domain.accommodation.repository.AccommodationRepository;
 import com.sparta.quokkatravel.domain.admin.coupon.dto.CouponToUserReq;
 import com.sparta.quokkatravel.domain.admin.coupon.dto.CouponToUserRes;
 import com.sparta.quokkatravel.domain.admin.coupon.repository.AdminCouponRepository;
 import com.sparta.quokkatravel.domain.common.exception.NotFoundException;
+import com.sparta.quokkatravel.domain.common.rabbitmq.dto.MessageRes;
+import com.sparta.quokkatravel.domain.common.rabbitmq.util.RabbitMqProducerUtil;
 import com.sparta.quokkatravel.domain.coupon.entity.Coupon;
 import com.sparta.quokkatravel.domain.coupon.entity.CouponType;
 import com.sparta.quokkatravel.domain.coupon.repository.CouponRepository;
@@ -16,7 +16,6 @@ import com.sparta.quokkatravel.domain.event.repository.EventRepository;
 import com.sparta.quokkatravel.domain.user.entity.User;
 import com.sparta.quokkatravel.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -27,12 +26,12 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class CouponProducerService {
 
-    private final RabbitTemplate rabbitTemplate;
     private final UserRepository userRepository;
     private final CouponRepository couponRepository;
     private final AccommodationRepository accommodationRepository;
     private final EventRepository eventRepository;
     private final AdminCouponRepository adminCouponRepository;
+    private final RabbitMqProducerUtil rabbitMqProducerUtil;
 
     /**
      * [2] Direct Exchange Binding
@@ -53,17 +52,10 @@ public class CouponProducerService {
             couponCode = UUID.randomUUID().toString().replace("-", "").substring(0, 12);
         } while (couponRepository.findByCode(couponCode).isPresent());
 
-        // RabbitMQ - Direct Exchange Binding
-        try {
-            // 1. 전송하려는 객체를 문자열로 변환
-            ObjectMapper objectMapper = new ObjectMapper();
-            String objectToJSON = objectMapper.writeValueAsString(couponMessageReq);
+        MessageRes messageRes = new MessageRes(couponMessageReq.getCouponName(), couponMessageReq.getCouponContent());
 
-            // 2. Direct Exchange 를 이용하여 routingKey 를 기반으로 queue2 로 데이터를 전송
-            rabbitTemplate.convertAndSend("coupon-issue-exchange","coupon.key", objectToJSON);
-        } catch (JsonProcessingException jpe) {
-            System.out.println("Parsing error: " + jpe.getMessage());
-        }
+        // RabbitMQ - Direct Exchange Binding
+        rabbitMqProducerUtil.sendMessageToQueue("coupon-issue-exchange", "coupon.key", messageRes);
 
         Coupon coupon = new Coupon(
                 couponMessageReq.getCouponName(),
