--- src/main/java/com/sparta/quokkatravel/domain/admin/coupon/util/AdminCouponUtil.java
@@ -1,4 +1,0 @@
-package com.sparta.quokkatravel.domain.admin.coupon.util;
-
-public class AdminCouponUtil {
-}
