+++ src/main/java/com/sparta/quokkatravel/domain/common/rabbitmq/service/ProducerService.java
@@ -1,7 +1,7 @@
-package com.sparta.quokkatravel.domain.rabbitmq.service;
+package com.sparta.quokkatravel.domain.common.rabbitmq.service;
 
+import com.sparta.quokkatravel.domain.common.rabbitmq.dto.MessageRes;
 import com.sparta.quokkatravel.domain.coupon.dto.request.MessageRequestDto;
-import com.sparta.quokkatravel.domain.rabbitmq.dto.MessageRes;
 
 public interface ProducerService {
